Added option to show previous nonempty frame of polygons in blue.
Hitting the 't' key now invokes a macro that toggles whether or not the polygons
in the previous nonempty frame are shown on the current frame in blue alongside
those in green/red of the current frame.  This is a piece in being able to edit
the assignment of predecessors to in mroi.