Wrapped a container class (RoiContainer) around Geometry so I can add more data.
All the commands are updated to use RoiContainer instead of Geometry, and several
of the functions in State.java have changed to work with this instead.  Also,
Save.java now writes out Lisp instead of the earlier format.  Load.java still
loads the old format, and will be kept as LegacyLoad.java in the next version
while a new Load.java will load the Lisp format.