Finished change to new file format (entirely Lisp based).
Wrote a parser in JavaCC and added an ant build target for it.  The new file
format is completely integrated.  There was a weird bug resulting from JavaCC
wanting to use static methods and having to call the .ReInit() method on the
parser each time.  The old file format is still supported from a macro in
the menu, but it no longer gets a keybinding.