Fixed compilation of ScmExport, and naming of measurements.
Measurements had to be in camelCase since atoms are not automatically mangled
to Java names when calling 'field' on an object.  The compilation problem
turned out to be a weird bug in Kawa where both of the following will work
in the interpreter:
(define (boolean-array)
 (make <boolean[]> length: 4))
(define (boolean-array)
 (apply make (cons <boolean[]> '(#f #f #f #f))))
but only the first will compile properly on the command line in Kawa 1.9.90.